feat: Add index and first recipe to website
This is the first commit containing actual content. The `index.html`
file is the landing page for the website, which contains a variety of
recipes.

This commit adds the first recipe (stuffed eggplant) to the respective
`recipes` directory. Additionally, an illustrative image of the dish is
added in the `image` directory.